+++ packages/openapi-parser/src/types/index.ts
@@ -1,4 +1,4 @@
-import type { OpenAPI } from '@scalar/openapi-types'
+import type { OpenAPI } from '@mintlify/openapi-types'
 
 import type { ERRORS, OpenApiVersion } from '../configuration/index.js'