+++ packages/openapi-parser/tests/benchmark/petstore.bench.ts
@@ -10,7 +10,7 @@ describe('petstore', () => {
     await resolveOld(specification)
   })
 
-  bench('@scalar/openapi-parser', async () => {
+  bench('@mintlify/openapi-parser', async () => {
     // Action!
     await resolveNew(specification)
   })