+++ scripts/bootstrap-package/vite.config.ts
@@ -1,4 +1,4 @@
-import { findEntryPoints } from '@scalar/build-tooling'
+import { findEntryPoints } from '@mintlify/build-tooling'
 import { URL, fileURLToPath } from 'node:url'
 import { defineConfig } from 'vite'